+++ delphi-frontend/src/main/java/au/com/integradev/delphi/symbol/SymbolTableBuilder.java
@@ -78,6 +78,7 @@ public class SymbolTableBuilder {
   private Path standardLibraryPath;
   private SearchPath searchPath = SearchPath.create(Collections.emptyList());
   private List<Path> sourceFiles = Collections.emptyList();
+  private List<Path> referencedFiles = Collections.emptyList();
   private Set<String> conditionalDefines = Collections.emptySet();
   private Set<String> unitScopeNames = Collections.emptySet();
   private Map<String, String> unitAliases = Collections.emptyMap();
@@ -97,6 +98,11 @@ public class SymbolTableBuilder {
 
   public SymbolTableBuilder sourceFiles(@NotNull List<Path> sourceFiles) {
     this.sourceFiles = sourceFiles;
+    return this;
+  }
+
+  public SymbolTableBuilder referencedFiles(@NotNull List<Path> referencedFiles) {
+    this.referencedFiles = referencedFiles;
     return this;
   }
 
@@ -447,6 +453,7 @@ public class SymbolTableBuilder {
 
     processStandardLibrarySearchPaths();
     searchPath.getRootDirectories().forEach(this::processSearchPath);
+    referencedFiles.forEach(file -> this.createUnitData(file, false));
     sourceFiles.forEach(file -> this.createUnitData(file, true));
 
     ProgressReport progressReport =